Emit more events and various improvements

+++ zou/app/services/news_service.py
@@ -46,10 +46,10 @@ def create_news_for_task_and_comment(task, comment, change=False):
         "news:new",
         {
             "news_id": news["id"],
-            "project_id": task["project_id"],
             "task_status_id": comment["task_status_id"],
             "task_type_id": task["task_type_id"],
         },
+        project_id=task["project_id"],
     )
     return news
 
@@ -60,9 +60,15 @@ def delete_news_for_comment(comment_id):
     comment afterwards.
     """
     news_list = News.get_all_by(comment_id=comment_id)
-    for news in news_list:
-        news.delete()
-        events.emit("news:delete", {"news_id": news.id})
+    if len(news_list) > 0:
+        task = tasks_service.get_task(news_list[0].task_id)
+        for news in news_list:
+            news.delete()
+            events.emit(
+                "news:delete",
+                {"news_id": news.id},
+                project_id=task["project_id"]
+            )
     return fields.serialize_list(news_list)
